backend/requirements.txt
blinker==1.8.2
click==8.1.7
Flask==2.2.5
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
Werkzeug==2.3.8
flask-restx==1.0.5
flask-cors==3.0.10
flask-login==0.6.2
scheduler
playwright
playwright-stealth